First working but non-complete implementation of TCP-Client and TCP-Server.
The application's UI has a client- and a server-tab that can be used to
switch between the two types of sockets.

The sockets inherit from QTcpSocket and QTcpServer, respectively.

The client socket is able to connect and disconnect to a TCP-Server.
The server socket is able to listen for incomming connections on a
specified port and automatically accepts them.

Data transfer is not implemented in the application, yet.