Reimplemented thread handling of ConnectedClients in ServerSocket regarding the following guide: https://mayaposch.wordpress.com/2011/11/01/how-to-really-truly-use-qthreads-the-full-explanation/
ConnectedClient no longer inherits from QThread (major error).
It now inherits from QObject and does not have a parent.
This way, the ConnectedClient object can be moved to an own thread that
is automatically killed when the client disconnects from the server.

The clientList now stores the objects of connected clients but not their
threads.