Seperate implementation of command- and data-connection.
The client and server now create an own socket for command- and
data-connection.
Both client and server now need the IP and two ports for connecting and
listening, respectively.

Static constant integers are defined in ClientSocket.h and
ConnectedClient.h to distinguish between the two types of connections in
signal and slot calls.

Client and server provide several signals and slots to communicate with
either the command- or data-connection seperately.

Note:
This code is still under test. Errors may occur!
Updated documentation will be commited as soon as possible.