Fixed wrong singal-emit in ServerSocket.
The signal newClient was emitted with the wring clientId.
Before this commit the position in the client-list was used as a signal-parameter.
That can be wrong if client disconnects from the server and a new one establishes
a connection. Now the cliendId will be taken directly from the ConnectedClient
object.